chore(gen): rename files properly
